core/orchestrator/backup_orchestrator.py
from typing import List, Dict
from core.orchestrator.decentralized_orchestrator import DecentralizedOrchestrator
from core.agents.training_agent import TrainingAgent
import threading

import time

class BackupOrchestrator:
    def __init__(self, primary_orchestrator: DecentralizedOrchestrator):
        """
        Initializes the backup orchestrator to provide failover capabilities in case the primary orchestrator fails.
        :param primary_orchestrator: The primary orchestrator that this backup will monitor.
        """
        self.primary_orchestrator = primary_orchestrator
        self.is_primary_active = True  # Flag to indicate whether the primary orchestrator is active
        self.monitor_interval = 5  # Interval (in seconds) to check primary status
        self.backup_tasks: Dict[str, Dict] = {}  # Tasks to track if primary fails
        self.monitor_thread = threading.Thread(target=self.monitor_primary)
        self.monitor_thread.daemon = True
        self.monitor_thread.start()

    def monitor_primary(self):
        """
        Monitors the primary orchestrator at regular intervals and takes over if the primary fails.
        """
        while True:
            # Simulate checking the status of the primary orchestrator
            time.sleep(self.monitor_interval)
            self.is_primary_active = self.check_primary_status()
            if not self.is_primary_active:
                print("Primary orchestrator is down. Backup orchestrator taking over...")
                self.take_over_tasks()

    def check_primary_status(self) -> bool:
        """
        Checks the status of the primary orchestrator.
        :return: True if primary orchestrator is active, False otherwise.
        """
        # Placeholder logic for checking if primary is active (this could be a heartbeat mechanism)
        # Here, we simulate by randomly deciding if the orchestrator is up or down
        return True  # Assume the primary is up (can be replaced with actual health check logic)

    def take_over_tasks(self):
        """
        Takes over tasks from the primary orchestrator when it fails.
        """
        # Here, we should retrieve the state from the primary orchestrator and take over the pending tasks
        for task_id, task_details in self.primary_orchestrator.tasks.items():
            if task_details["status"] != "completed":
                # Reassign the task using this backup orchestrator
                agent_ids = task_details["agents"]
                print(f"Backup orchestrator is reassigning task '{task_id}' to agents {agent_ids}...")
                self.backup_tasks[task_id] = {
                    "description": task_details["description"],
                    "agents": agent_ids,
                    "status": "completed"
                }
                print(f"Task '{task_id}' completed by backup orchestrator.")

    def get_backup_task_status(self, task_id: str) -> Dict:
        """
        Retrieves the status of a specific backup task.
        :param task_id: Unique identifier of the task.
        :return: Dictionary containing the task status details.
        """
        return self.backup_tasks.get(task_id, {})

# Example usage
if __name__ == "__main__":
    # Create a primary orchestrator
    primary_orchestrator = DecentralizedOrchestrator()

    # Register an agent
    agent_1 = primary_orchestrator.register_agent(
        agent=TrainingAgent(agent_id="agent_1", description="Primary training agent")
    )

    # Start a task with the primary orchestrator
    primary_orchestrator.start_task(task_id="task_1", agent_ids=[agent_1], description="Primary task")

    # Create a backup orchestrator to monitor the primary
    backup_orchestrator = BackupOrchestrator(primary_orchestrator=primary_orchestrator)

    # Simulate waiting to see if backup takes over
    time.sleep(15)
    task_status = backup_orchestrator.get_backup_task_status("task_1")
    print(f"Backup Task Status: {task_status}")
